+++ website_v2/package.json
@@ -28,8 +28,8 @@
     "postcss": "8.4.23",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-linkedin-login-oauth2": "^2.0.1",
     "react-redux": "^8.0.7",
+    "reactjs-social-login": "^2.6.3",
     "recharts": "^2.7.2",
     "tailwindcss": "3.3.2",
     "typescript": "5.0.4",
