--- website_v2/src/app/linkedin/page.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import {LinkedInCallback} from "react-linkedin-login-oauth2";
-
-// Named export can't be used here as the library is using the default export internally
-export default function LinkedInPage() {
-  return <LinkedInCallback />;
-}
