+++ website_v2/src/components/landing/hero_main.tsx
@@ -1,8 +1,12 @@
 "use client";
 
 import {useSession} from "next-auth/react";
-import {Linkedin} from "./linkedin_main";
 import Image from "next/image";
+import dynamic from "next/dynamic";
+// Dynamically load non SSR component
+const LinkedIn = dynamic(() => import("./linkedin").then(mod => mod.LinkedIn), {
+  ssr: false,
+});
 
 const Hero = () => {
   const {status} = useSession();
@@ -24,7 +28,7 @@ const Hero = () => {
           {/* Buttons */}
           {status === "unauthenticated" ? (
             <div className="mt-4 w-full flex flex-col justify-center items-center gap-2 md:flex-row md:justify-start md:gap-4">
-              <Linkedin />
+              <LinkedIn />
             </div>
           ) : null}
         </div>
